+++ src/elements/Table/TableRow.jsx
@@ -17,6 +17,7 @@ import classNames from 'classnames';
         return (
             <TableRow
                 multiline={false} // boolean true or false, allow linebreaks and multilined content in table cells (optional, default: false)
+                preventCellOverflow={true} // boolean true or false, prevent overflowing content in table cells (optional, default: false)
                 className="my-own-class" // string, used for additional CSS class descriptions
             >
                 <!-- content -->
@@ -29,10 +30,17 @@ import classNames from 'classnames';
  */
 
 const TableRow = props => {
-    const {className, children, multiline, ...otherProps} = props;
+    const {
+        className,
+        children,
+        multiline,
+        preventCellOverflow,
+        ...otherProps
+    } = props;
     const rowClassNames = classNames(
         {
             'mdl-data-table--multiline': multiline === true,
+            'mdl-data-table--preventoverflow': preventCellOverflow === true,
         },
         className
     );
@@ -53,11 +61,16 @@ TableRow.propTypes = {
      * allow linebreaks and multilined content in table cells
      */
     multiline: Proptypes.bool,
+    /**
+     * prevent overflowing content in table cells
+     */
+    preventCellOverflow: Proptypes.bool,
 };
 
 TableRow.defaultProps = {
     className: '',
     multiline: false,
+    preventCellOverflow: false,
 };
 
 TableRow.displayName = 'Table row';