+++ CHANGELOG.md
@@ -4,7 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Ch
 
 ## [Unreleased]
 
-TODO: add at least one Added, Changed, Deprecated, Removed, Fixed or Security section
+### Added
+- `scrollTableOverflow` option for `<Table />`, this is now not enabled by default
+- `preventCellOverflow` option for `<Table />`, this is now not enabled by default
 
 ## [3.0.0] 2018-10-23